CatalystSocket Token API

+++ services/network/catalyst/catalyst_socket.h
@@ -93,12 +93,12 @@ class COMPONENT_EXPORT(NETWORK_SERVICE) CatalystSocket : public mojom::CatalystS
 
     static const uint32_t kMaxReadSize = 65535;
 
-    static const uint32_t kSegmentSize = 1460;
-    static const uint32_t kStartCwndSize = 10 * kSegmentSize;
+    static const uint64_t kSegmentSize = 1460;
+    static const uint64_t kStartCwndSize = 10 * kSegmentSize;
     static constexpr float kBeta = 0.6;
     static constexpr float kAlpha = (3.0  * (kBeta / (2.0 - kBeta)));
     static const uint64_t kNumRTTs = 12;
-    static const uint32_t kProbeSizeBytes = 2; 
+    static const uint64_t kProbeSizeBytes = 2; 
     static const uint64_t kStartRTTns = 200 * 1e6;
     static const uint64_t kRTTFactorTimeout = 2;
 
@@ -119,7 +119,7 @@ class COMPONENT_EXPORT(NETWORK_SERVICE) CatalystSocket : public mojom::CatalystS
 
     int CwndAvailable();
     void Loss(int num_losses);
-    void Ack(uint32_t packet_size);
+    void Ack(uint64_t packet_size);
     uint64_t RTT();
     uint64_t Timeout();
     void OnRTTTimer();
@@ -151,11 +151,11 @@ class COMPONENT_EXPORT(NETWORK_SERVICE) CatalystSocket : public mojom::CatalystS
 
     float ssthresh_ = 65536 * 1.90;
     unsigned char phase_ = kPhaseSlowStart;
-    uint32_t cwnd_size_ = kStartCwndSize;
-    uint32_t cwnd_used_ = 0;
+    uint64_t cwnd_size_ = kStartCwndSize;
+    uint64_t cwnd_used_ = 0;
     uint16_t last_seq_num_ = 1;
     std::map<uint16_t, std::chrono::time_point<std::chrono::steady_clock>> unacked_sent_at_;
-    std::map<uint16_t, uint32_t> unacked_sizes_;
+    std::map<uint16_t, uint64_t> unacked_sizes_;
     std::set<std::uint16_t> unacked_;
     uint64_t rtts_[kNumRTTs];
     uint64_t rtt_index_ = 0;
